Install Electron
I installed the electron framework using Yarn. I then was able to get the project to give us proof of life. Implementing us the versioning that we are using in the window, I also configured the closing and the opening of windows for all 3 OS platforms. We should be able to continue development here.