isPlaying
I added a new API to see if the video is playing or not. This will allow the iOS app to have a better knowledge on the state of what is currently happening on the TV. There is going to need to be some improvements to ensure that all of this stuff works.