Indefinite Pause
I fixed the polling issue on the pause button. Now, if a user does truly click the pause button it will stay paused until they reinstate the play button. Previously the constant polling for the m3u file would cause a bug with it trying to find the updated version of the file. However, once the m3u is now found it will quit polling.